change name - from forecast to forecaster - from anomaly to detector/anomaly
