+++ docs/readthedocs/source/doc/PythonAPI/Chronos/forecasters.rst
@@ -10,7 +10,7 @@ Long short-term memory(LSTM) is a special type of recurrent neural network(RNN).
 
 For the detailed algorithm description, please refer to `here <https://github.com/intel-analytics/analytics-zoo/blob/master/docs/docs/Chronos/Algorithm/LSTMAlgorithm.md>`__.
 
-.. automodule:: zoo.chronos.forecast.lstm_forecaster
+.. automodule:: zoo.chronos.forecaster.lstm_forecaster
     :members:
     :undoc-members:
     :show-inheritance:
@@ -23,7 +23,7 @@ Please refer to BasePytorchForecaster for other methods other than initializatio
 
 Seq2SeqForecaster wraps a sequence to sequence model based on LSTM, and is suitable for multivariant & multistep time series forecasting.
 
-.. automodule:: zoo.chronos.forecast.seq2seq_forecaster
+.. automodule:: zoo.chronos.forecaster.seq2seq_forecaster
     :members:
     :undoc-members:
     :show-inheritance:
@@ -36,7 +36,7 @@ Please refer to BasePytorchForecaster for other methods other than initializatio
 
 Temporal Convolutional Networks (TCN) is a neural network that use convolutional architecture rather than recurrent networks. It supports multi-step and multi-variant cases. Causal Convolutions enables large scale parallel computing which makes TCN has less inference time than RNN based model such as LSTM.
 
-.. automodule:: zoo.chronos.forecast.tcn_forecaster
+.. automodule:: zoo.chronos.forecaster.tcn_forecaster
     :members:
     :undoc-members:
     :show-inheritance:
@@ -57,7 +57,7 @@ TCMFForecaster supports distributed training and inference. It is based on Orca 
 * You can refer to `TCMFForecaster installation <https://github.com/intel-analytics/analytics-zoo/blob/master/docs/docs/Chronos/tutorials/TCMFForecaster.md/#step-0-prepare-environment>`__ to install required packages.
 * Your operating system (OS) is required to be one of the following 64-bit systems: **Ubuntu 16.04 or later** and **macOS 10.12.6 or later**.
 
-.. automodule:: zoo.chronos.forecast.tcmf_forecaster
+.. automodule:: zoo.chronos.forecaster.tcmf_forecaster
     :members:
     :undoc-members:
     :show-inheritance:
@@ -72,7 +72,7 @@ MTNet is proposed by paper `A Memory-Network Based Solution for Multivariate Tim
 
 For the detailed algorithm description, please refer to `here <https://github.com/intel-analytics/analytics-zoo/blob/master/docs/docs/Chronos/Algorithm/MTNetAlgorithm.md>`__.
 
-.. automodule:: zoo.chronos.forecast.mtnet_forecaster
+.. automodule:: zoo.chronos.forecaster.mtnet_forecaster
     :members:
     :undoc-members:
     :show-inheritance:
@@ -83,7 +83,7 @@ ARIMAForecaster
 
 AutoRegressive Integrated Moving Average (ARIMA) is a class of statistical models for analyzing and forecasting time series data. It consists of 3 components: AR (AutoRegressive), I (Integrated) and MA (Moving Average). In ARIMAForecaster we use the SARIMA model (Seasonal ARIMA), which is an extension of ARIMA that additionally supports the direct modeling of the seasonal component of the time series.
 
-.. automodule:: zoo.chronos.forecast.arima_forecaster
+.. automodule:: zoo.chronos.forecaster.arima_forecaster
     :members:
     :undoc-members:
     :show-inheritance:
@@ -96,7 +96,7 @@ Prophet is a procedure for forecasting time series data based on an additive mod
 
 For the detailed algorithm description, please refer to `here <https://github.com/facebook/prophet>`__.
 
-.. automodule:: zoo.chronos.forecast.prophet_forecaster
+.. automodule:: zoo.chronos.forecaster.prophet_forecaster
     :members:
     :undoc-members:
     :show-inheritance:
@@ -105,7 +105,7 @@ For the detailed algorithm description, please refer to `here <https://github.co
 chronos.forecast.tfpark_forecaster
 ----------------------------------------
 
-.. automodule:: zoo.chronos.forecast.tfpark_forecaster
+.. automodule:: zoo.chronos.forecaster.tfpark_forecaster
     :members:
     :undoc-members:
     :show-inheritance:
@@ -113,6 +113,6 @@ chronos.forecast.tfpark_forecaster
 chronos.forecast.base_forecaster.BasePytorchForecaster
 ----------------------------------------
 
-.. autoclass:: zoo.chronos.forecast.base_forecaster.BasePytorchForecaster
+.. autoclass:: zoo.chronos.forecaster.base_forecaster.BasePytorchForecaster
     :members:
     :show-inheritance:
