+++ pyzoo/test/zoo/chronos/detector/anomaly/test_dbscan_detector.py
@@ -18,7 +18,7 @@ import pytest
 import numpy as np
 from test.zoo.pipeline.utils.test_utils import ZooTestCase
 
-from zoo.chronos.anomaly.dbscan_detector import DBScanDetector
+from zoo.chronos.detector.anomaly.dbscan_detector import DBScanDetector
 
 
 class TestDBScanDetector(ZooTestCase):
