+++ pyzoo/test/zoo/chronos/forecaster/test_mtnet_forecaster.py
@@ -20,7 +20,7 @@ from zoo.chronos.autots.deprecated.feature.time_sequence import TimeSequenceFeat
 import tensorflow as tf
 import pandas as pd
 
-from zoo.chronos.forecast.mtnet_forecaster import MTNetForecaster
+from zoo.chronos.forecaster.mtnet_forecaster import MTNetForecaster
 from unittest import TestCase
