+++ pyzoo/test/zoo/chronos/forecaster/test_tcmf_forecaster.py
@@ -16,7 +16,7 @@
 
 import pytest
 import numpy as np
-from zoo.chronos.forecast.tcmf_forecaster import TCMFForecaster
+from zoo.chronos.forecaster.tcmf_forecaster import TCMFForecaster
 from unittest import TestCase
 import tempfile
 import pandas as pd
