+++ pyzoo/zoo/chronos/forecaster/arima_forecaster.py
@@ -14,7 +14,7 @@
 # limitations under the License.
 #
 
-from zoo.chronos.forecast.abstract import Forecaster
+from zoo.chronos.forecaster.abstract import Forecaster
 from zoo.chronos.model.arima import ARIMAModel
