+++ pyzoo/zoo/chronos/forecaster/tcn_forecaster.py
@@ -14,8 +14,8 @@
 # limitations under the License.
 #
 
-from zoo.chronos.forecast.base_forecaster import BasePytorchForecaster
-from zoo.chronos.forecast.utils import set_pytorch_seed
+from zoo.chronos.forecaster.base_forecaster import BasePytorchForecaster
+from zoo.chronos.forecaster.utils import set_pytorch_seed
 from zoo.chronos.model.tcn import TCNPytorch
 from zoo.chronos.model.tcn import model_creator, optimizer_creator, loss_creator
